Allow reading of headerless compressed pages
Summary:
WebScaleSQL Feature: 2012 InnoDB Compression

The complete version of 2012 compression writes headerless zlib data
blobs.  This is true in both the 5.1 and 5.6 Facebook branches.

This change allows WebScaleSQL to read those pages, in addition to
those with headers.

But, this still writes only with zlib headers, not without.

A later diff will add support to WebScaleSQL for actually writing
these headerless pages.

Needed to backport inflateReset2() from a future version of zlib in
order to reset a decompression stream without allocating memory for its
internal state and with specifying windowBits. This was necessary to
automatically determine whether a compressed page was compressed using
adler32 checksums or not. There should not be any issues in terms of
backwards compatibility with this. After this, a clean build is required
for stress tests to pass.

Test Plan: Tested reading data compressed by the FB branch.  Worked.

Reviewers: pivanof, inaam-rana

Reviewed By: inaam-rana

Subscribers: WebScaleSQL, jtolmer

Differential Revision: https://reviews.facebook.net/D17691